fix(OrgManager): org manager now features the same branding settings as Campus Settings
